requirements.txt
# Voice Recognition
SpeechRecognition>=3.10.0
pyttsx3==2.90
PyAudio>=0.2.13

# System Control
pyautogui==0.9.54
psutil>=5.9.0
keyboard==0.13.5
mouse==0.7.1

# GUI
customtkinter>=5.2.0
Pillow>=10.0.0

# AI and Natural Language Processing
openai>=1.0.0
google-generativeai>=0.3.0
nltk>=3.8.1

# Utility
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.26.0
aiohttp==3.9.1
asyncio==3.4.3

# Alternative Voice Recognition
vosk==0.3.45
